Add files via upload
installed the Redux library along with React-Redux, a library to help integrate Redux with React applications. Set up a reducer and used it to initialize the Redux store state with the createStore() function from Redux.  then used the <Provider> component from Redux to make the store available to child components connected with the connect() function from React-Redux. Then connected the Main component with connect() and made the store state available to the Main component as props using mapStateToProps()